Consolidate app banners into AppBanner (DEV-2339)
Move TrialBanner, DevModeBanner, ReadOnlyBanner, PaymentBanner, and
LicenseTokenMissingBanner from nav/components/ into
common/components/AppBanner/ - AppBanner.tsx is their only consumer, and
it lives in common while the banners lived in nav.

Verified via dependency-graph analysis + repo-wide greps (incl. e2e).
Pure git mv + import updates; no behavior change.
