+++ frontend/src/metabase/common/components/AppBanner/LicenseTokenMissingBanner/LicenseTokenMissingBanner.module.css
