+++ frontend/src/metabase/common/components/AppBanner/LicenseTokenMissingBanner/index.ts
