+++ frontend/src/metabase/common/components/AppBanner/ReadOnlyBanner/ReadOnlyBanner.tsx
