+++ frontend/src/metabase/common/components/AppBanner/ReadOnlyBanner/index.ts
