+++ frontend/src/metabase/common/components/AppBanner/TrialBanner/TrialBanner.tsx
