+++ frontend/src/metabase/common/components/AppBanner/TrialBanner/TrialBanner.unit.spec.tsx
