+++ frontend/src/metabase/common/components/AppBanner/TrialBanner/index.ts
